+++ education_app/lib/Pages/LandingPage.dart
@@ -1,10 +1,9 @@
-import 'package:education_app/Pages/QuizPages/HistoryPages/AllQuizzes.dart';
-import 'package:education_app/Pages/QuizBuilder.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:education_app/Pages/AuthenticationPages/LoginPage.dart';
-import 'package:education_app/Pages/QuizPages/QuizPage.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+
+import 'package:education_app/Pages/QuizPages/HistoryPages/AllQuizzes.dart';
+import 'package:education_app/Pages/QuizBuilder.dart';
 import 'package:education_app/Pages/QuizPages/HistoryPages/QuizSummaryPage.dart';
 import 'package:education_app/Quizzes/quiz.dart';
 import 'package:education_app/Quizzes/quizManager.dart';
@@ -17,6 +16,10 @@ class LandingPage extends StatefulWidget {
 
 class _LandingPageState extends State<LandingPage> {
   User? _user;
+  List<String> userInterests = [];
+  int xpLevel = 0; // Assuming XP level is an integer
+  late String _displayName = "Placeholder";
+
   List<String> recentQuizzes = [];
   late List<QuizQuestion> loadedQuestions = [];
   Map<String, dynamic> quizAttemptData = {};
@@ -25,6 +28,15 @@ class _LandingPageState extends State<LandingPage> {
   String quizName = "";
   late Quiz quiz;
 
+  List<String> otherTopics = [
+    'Topic 1',
+    'Topic 2',
+    'Topic 3',
+    'Topic 4',
+    // Add more topics as needed
+  ];
+
+
   @override
   void initState() {
     super.initState();
@@ -32,22 +44,65 @@ class _LandingPageState extends State<LandingPage> {
     quizManager = QuizManager();
   }
 
-  @override
-  void dispose() {
-    super.dispose();
-  }
-
   void _checkAuthState() {
     FirebaseAuth.instance.authStateChanges().listen((User? user) {
       if (mounted) {
         setState(() {
-          _user = user; // Set the current user
+          _user = user;
         });
         if (user != null) {
+          _getUserInterests(user.uid);
+          _getUserXPLevel(user.uid);
+          _getUserDisplayName(user.uid); // Call to get user display name
           _checkQuizHistory();
         }
       }
     });
+  }
+
+  void _getUserInterests(String uid) async {
+    try {
+      DocumentSnapshot userSnapshot =
+          await FirebaseFirestore.instance.collection('users').doc(uid).get();
+
+      if (userSnapshot.exists) {
+        setState(() {
+          userInterests = List<String>.from(userSnapshot.get('interests'));
+        });
+      }
+    } catch (e) {
+      print('Error fetching user interests: $e');
+    }
+  }
+
+  void _getUserXPLevel(String uid) async {
+    try {
+      DocumentSnapshot userSnapshot =
+          await FirebaseFirestore.instance.collection('users').doc(uid).get();
+
+      if (userSnapshot.exists) {
+        setState(() {
+          xpLevel = userSnapshot.get('xpLvl');
+        });
+      }
+    } catch (e) {
+      print('Error fetching user XP level: $e');
+    }
+  }
+
+  void _getUserDisplayName(String uid) async {
+    try {
+      DocumentSnapshot userSnapshot =
+          await FirebaseFirestore.instance.collection('users').doc(uid).get();
+
+      if (userSnapshot.exists) {
+        setState(() {
+          _displayName = userSnapshot.get('displayName');
+        });
+      }
+    } catch (e) {
+      print('Error fetching user display name: $e');
+    }
   }
 
   Future<List<String>> getQuizNames(List<String> quizIds) async {
@@ -182,335 +237,6 @@ class _LandingPageState extends State<LandingPage> {
     };
   }
 
-  @override
-  Widget build(BuildContext context) {
-    double screenWidth = MediaQuery.of(context).size.width;
-
-    return Scaffold(
-      appBar: AppBar(
-        automaticallyImplyLeading: false,
-        title: Center(child: const Text('IntelliQuiz')),
-        actions: [
-          Center(
-            child: IconButton(
-              icon: const Icon(Icons.logout),
-              onPressed: () async {
-                await FirebaseAuth.instance.signOut();
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => LoginPage(),
-                  ),
-                );
-              },
-            ),
-          ),
-        ],
-      ),
-      body: Center(
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: <Widget>[
-            Container(
-              margin: const EdgeInsets.all(30.0),
-              width: MediaQuery.of(context).size.width * 0.4,
-              child: Column(
-                children: [
-                  Container(
-                    alignment: Alignment.center,
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(12),
-                      color: const Color(0xFFf3edf6).withOpacity(1),
-                      boxShadow: [
-                        BoxShadow(
-                          color: Colors.black.withOpacity(0.2),
-                          spreadRadius: 5,
-                          blurRadius: 10,
-                          offset: const Offset(0, 3),
-                        ),
-                      ],
-                    ),
-                    width: MediaQuery.of(context).size.width * 0.5,
-                    height: MediaQuery.of(context).size.height * 0.1, // Controls the height of the top left container
-                    margin: const EdgeInsets.fromLTRB(0.0, 0.0, 0.0, 20.0),
-                    child: FutureBuilder<String?>(
-                    future: getUserDisplayName(_user!.uid),
-                    builder: (context, snapshot) {
-                      if (snapshot.connectionState == ConnectionState.waiting) {
-                        return Text(
-                          'Welcome, Loading...',
-                          style: const TextStyle(fontSize: 20),
-                        );
-                      } else if (snapshot.hasError) {
-                        return Text(
-                          'Error loading display name',
-                          style: const TextStyle(fontSize: 20),
-                        );
-                      } else {
-                        String? displayName = snapshot.data;
-
-                        return Text(
-                          'Welcome, ${displayName ?? 'User'}!',
-                          style: const TextStyle(fontSize: 20),
-                        );
-                      }
-                    },
-                  ),
-
-                  ),
-                  Container(
-                    alignment: Alignment.topCenter,
-                    width: MediaQuery.of(context).size.width * 0.5,
-                    height: MediaQuery.of(context).size.height * 0.75, // Controls the height of the bottom left container
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(12),
-                      color: const Color(0xFFf3edf6).withOpacity(1),
-                      boxShadow: [
-                        BoxShadow(
-                          color: Colors.black.withOpacity(0.2),
-                          spreadRadius: 5,
-                          blurRadius: 10,
-                          offset: const Offset(0, 3),
-                        ),
-                      ],
-                    ),
-                    child: Column(
-                      children: [
-                        Container(
-                          alignment: Alignment.center,
-                          margin: const EdgeInsets.fromLTRB(0.0, 20.0, 0.0, 10.0),
-                          child: Text('Available', style: GoogleFonts.nunito(color: Colors.black, fontSize: 28)),
-                        ),
-                        Text('Test yourself!', style: GoogleFonts.nunito(color: Colors.black, fontSize: 20)),
-                        Expanded(
-                          child: QuizListView(), // quizListView widget to display quizzes
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-            Container(
-              margin: const EdgeInsets.all(30.0),
-              width: MediaQuery.of(context).size.width * 0.4,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.stretch,
-                children: [
-                  Container(
-                    alignment: Alignment.center,
-                    height: MediaQuery.of(context).size.height * 0.1, // Controls the height of the top right container
-                    margin: const EdgeInsets.only(bottom: 20.0),
-                    decoration: BoxDecoration(
-                      borderRadius: BorderRadius.circular(12),
-                      color: const Color(0xFFf3edf6).withOpacity(1),
-                      boxShadow: [
-                        BoxShadow(
-                          color: Colors.black.withOpacity(0.2),
-                          spreadRadius: 5,
-                          blurRadius: 10,
-                          offset: const Offset(0, 3),
-                        ),
-                      ],
-                    ),
-                    child: Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        ElevatedButton(
-                          onPressed: () {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                builder: (context) => QuizPage(),
-                              ),
-                            );
-                          },
-                          child: const Text(
-                            'Take Quiz',
-                            style: TextStyle(color: Colors.black),
-                          ),
-                        ),
-                        const SizedBox(width: 20),
-                        ElevatedButton(
-                          onPressed: () {
-                            Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                builder: (context) => const QuizBuilder(),
-                              ),
-                            );
-                          },
-                          child: const Text(
-                            'Quiz Builder',
-                            style: TextStyle(color: Colors.black),
-                          ),
-                        ),
-                        const SizedBox(width: 20),
-                        ElevatedButton(
-                          onPressed: () {
-                            _checkQuizHistory();
-                          },
-                          child: const Text(
-                            'Check Quiz History',
-                            style: TextStyle(color: Colors.black),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  Expanded(
-                    child: Container(
-                      decoration: BoxDecoration(
-                        borderRadius: BorderRadius.circular(12),
-                        color: const Color(0xFFf3edf6).withOpacity(1),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.black.withOpacity(0.2),
-                            spreadRadius: 5,
-                            blurRadius: 10,
-                            offset: const Offset(0, 3),
-                          ),
-                        ],
-                      ),
-                      child: SingleChildScrollView(
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            const SizedBox(height: 20),
-                            Text(
-                              'History',
-                              style: GoogleFonts.nunito(color: Colors.black, fontSize: 28),
-                            ),
-                            Text(
-                              'View your recent efforts!',
-                              style: GoogleFonts.nunito(color: Colors.black, fontSize: 20),
-                            ),
-                            const SizedBox(height: 20),
-                            FutureBuilder<List<String>>(
-                              future: getQuizNames(recentQuizzes),
-                              builder: (context, snapshot) {
-                                if (snapshot.connectionState == ConnectionState.waiting) {
-                                  return Center(child: CircularProgressIndicator());
-                                } else if (snapshot.hasError) {
-                                  return Center(child: Text('Error loading quiz names'));
-                                } else {
-                                  List<String> quizNames = snapshot.data ?? [];
-                                  int numRecentQuizzes = quizNames.length;
-                                  int numQuizzesPerRow = 2;
-                                  int numRows = (numRecentQuizzes / numQuizzesPerRow).ceil();
-                                  List<Widget> rows = List.generate(numRows, (rowIndex) {
-                                    List<Widget> rowChildren = [];
-                                    for (int i = 0; i < numQuizzesPerRow; i++) {
-                                      int index = rowIndex * numQuizzesPerRow + i;
-                                      if (index < numRecentQuizzes) {
-                                        rowChildren.add(
-                                          Expanded(
-                                            child: Padding(
-                                              padding: const EdgeInsets.symmetric(horizontal: 10.0),
-                                              child: Container(
-                                                decoration: BoxDecoration(
-                                                  borderRadius: BorderRadius.circular(10),
-                                                  color: Colors.white,
-                                                  boxShadow: [
-                                                    BoxShadow(
-                                                      color: Colors.black.withOpacity(0.2),
-                                                      spreadRadius: 1,
-                                                      blurRadius: 2,
-                                                      offset: const Offset(0, 1),
-                                                    ),
-                                                  ],
-                                                ),
-                                                child: InkWell(
-                                                  onTap: () async {
-                                                    await _getloadedQuestions(recentQuizzes[index]);
-                                                    await _loadQuizAttemptData(recentQuizzes[index]);
-                                                    _quizSummaryButton(loadedQuestions, quizAttemptData);
-                                                  },
-                                                  child: Column(
-                                                    crossAxisAlignment: CrossAxisAlignment.center,
-                                                    children: [
-                                                      const Icon(
-                                                        Icons.history,
-                                                        size: 60,
-                                                        color: Colors.blue,
-                                                      ),
-                                                      const SizedBox(height: 10),
-                                                      Text(
-                                                        'Recent Quiz ${index + 1}: ${quizNames[index]}',
-                                                        style: const TextStyle(fontSize: 16),
-                                                        textAlign: TextAlign.center,
-                                                      ),
-                                                    ],
-                                                  ),
-                                                ),
-                                              ),
-                                            ),
-                                          ),
-                                        );
-                                      }
-                                    }
-                                    return Row(
-                                      children: rowChildren,
-                                    );
-                                  });
-                                  return Column(
-                                    crossAxisAlignment: CrossAxisAlignment.start,
-                                    children: rows,
-                                  );
-                                }
-                              },
-                            ),
-                            const SizedBox(height: 20),
-                            Align(
-                              alignment: Alignment.topRight,
-                              child: Padding(
-                                padding: const EdgeInsets.only(top: 0.0, right: 20.0),
-                                child: ElevatedButton(
-                                  onPressed: () {
-                                    Navigator.push(
-                                      context,
-                                      MaterialPageRoute(
-                                        builder: (context) => QuizHistoryPage(),
-                                      ),
-                                    );
-                                  },
-                                  child: const Text(
-                                    'View All',
-                                    style: TextStyle(color: Colors.black),
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ]
-        )
-      )
-    );
-  }
-
-  Future<String?> getUserDisplayName(String userId) async {
-    try {
-      DocumentSnapshot userSnapshot = await FirebaseFirestore.instance.collection('users').doc(userId).get();
-
-      if (userSnapshot.exists) {
-        return userSnapshot.get('displayName');
-      } else {
-        print('User not found in Firestore');
-        return null;
-      }
-    } catch (e) {
-      print('Error retrieving user display name: $e');
-      return null;
-    }
-  } 
-
   void _quizSummaryButton(loadedQuestions, quizData) {
     Navigator.push(
       context,
@@ -522,76 +248,465 @@ class _LandingPageState extends State<LandingPage> {
       ),
     );
   }
-}
-
-class QuizListView extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return GridView.builder(
-      padding: const EdgeInsets.all(10.0), // Add padding to move tiles away from the container border
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 2,
-        crossAxisSpacing: 10.0,
-        mainAxisSpacing: 10.0,
-      ),
-      itemCount: 8, // Maximum of 8 tiles (4 high, 2 wide)
-      itemBuilder: (context, index) {
-        return AvailableInterestTile(index: index);
-      },
-    );
-  }
-}
-
-class AvailableInterestTile extends StatelessWidget {
-  final int index;
-
-  const AvailableInterestTile({required this.index});
 
   @override
   Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        // Placeholder for button functionality
-        print('Tile ${index + 1} tapped!');
-        // Add code here to generate quiz based on user's interest
-      },
-      child: Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(10),
-          color: Colors.white,
-          boxShadow: [
-            BoxShadow(
-              color: Colors.black.withOpacity(0.2),
-              spreadRadius: 1,
-              blurRadius: 2,
-              offset: const Offset(0, 1),
-            ),
-          ],
-        ),
-        // height: MediaQuery.of(context).size.height * 1, // Half the height of the screen
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const Icon(
-              Icons.history,
-              size: 40, // Reduce the icon size
-              color: Colors.blue,
-            ),
-            const SizedBox(height: 5), // Add some space between the icon and the text
-            Text(
-              'Interest ${index + 1}', // Placeholder for tile name
-              style: const TextStyle(fontSize: 14), // Reduce the font size
-              textAlign: TextAlign.center,
-            ),
-          ],
-        ),
+    return Scaffold(
+      appBar: AppBar(
+        automaticallyImplyLeading: false,
+        centerTitle: true,
+        title: Text('Quiz App'),
+        actions: _user != null
+            ? [
+                IconButton(
+                  icon: Icon(Icons.logout),
+                  onPressed: () async {
+                    await FirebaseAuth.instance.signOut();
+                  },
+                ),
+              ]
+            : null,
       ),
+      body: _user != null
+          ? Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Expanded(
+                  flex: 2,
+                  child: Container(
+                    height: MediaQuery.of(context).size.height,
+                    margin: const EdgeInsets.all(30.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Expanded(
+                          child: Container(
+                            width: MediaQuery.of(context).size.width * 2 / 3,
+                            padding: EdgeInsets.all(16.0),
+                            decoration: BoxDecoration(
+                              borderRadius: BorderRadius.circular(12),
+                              color: const Color(0xFFf3edf6).withOpacity(1),
+                              boxShadow: [
+                                BoxShadow(
+                                  color: Colors.black.withOpacity(0.2),
+                                  spreadRadius: 5,
+                                  blurRadius: 10,
+                                  offset: const Offset(0, 3),
+                                ),
+                              ],
+                            ),
+                            child: SingleChildScrollView(
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                children: [
+                                  const SizedBox(height: 20),
+                                  Text(
+                                    'Your Interests',
+                                    style: GoogleFonts.nunito(color: Colors.black, fontSize: 28),
+                                  ),
+                                  const SizedBox(height: 20),
+                                  FutureBuilder<List<String>>(
+                                    future: Future.value(userInterests),
+                                    builder: (context, snapshot) {
+                                      if (snapshot.connectionState == ConnectionState.waiting) {
+                                        return Center(child: CircularProgressIndicator());
+                                      } else if (snapshot.hasError) {
+                                        return Center(child: Text('Error loading interests'));
+                                      } else {
+                                        List<String> interests = snapshot.data ?? [];
+                                        int numInterests = interests.length;
+                                        int numInterestsPerRow = 4; // Adjust the number of interests per row as needed
+                                        int numRows = (numInterests / numInterestsPerRow).ceil();
+                                        List<Widget> rows = List.generate(numRows, (rowIndex) {
+                                          List<Widget> rowChildren = [];
+                                          for (int i = 0; i < numInterestsPerRow; i++) {
+                                            int index = rowIndex * numInterestsPerRow + i;
+                                            const SizedBox(height: 10);
+                                            if (index < numInterests) {
+                                              rowChildren.add(
+                                                Flexible(
+                                                  child: Padding(
+                                                    padding: const EdgeInsets.symmetric(horizontal: 10.0, vertical: 10),
+                                                    child: Container(
+                                                      height: 200,
+                                                      decoration: BoxDecoration(
+                                                        borderRadius: BorderRadius.circular(10),
+                                                        color: Colors.white,
+                                                        boxShadow: [
+                                                          BoxShadow(
+                                                            color: Colors.black.withOpacity(0.2),
+                                                            spreadRadius: 1,
+                                                            blurRadius: 2,
+                                                            offset: const Offset(0, 1),
+                                                          ),
+                                                        ],
+                                                      ),
+                                                      child: InkWell(
+                                                        onTap: () async {
+                                                          print('Interest ${index + 1}: ${interests[index]} pressed');
+                                                          // Add functionality here if needed
+                                                        },
+                                                        child: Center(
+                                                          child: Column(
+                                                            mainAxisAlignment: MainAxisAlignment.center,
+                                                            crossAxisAlignment: CrossAxisAlignment.center,
+                                                            children: [
+                                                              const Icon(
+                                                                Icons.history,
+                                                                size: 60,
+                                                                color: Colors.blue,
+                                                              ),
+                                                              const SizedBox(height: 10),
+                                                              Text(
+                                                                interests[index],
+                                                                style: const TextStyle(fontSize: 16),
+                                                                textAlign: TextAlign.center,
+                                                              ),
+                                                            ],
+                                                          ),
+                                                        ),
+                                                      ),
+                                                    ),
+                                                  ),
+                                                ),
+                                              );
+                                            } else {
+                                              rowChildren.add(Flexible(child: SizedBox())); // Add an empty Flexible widget for even distribution
+                                            }
+                                          }
+                                          return Row(
+                                            children: rowChildren,
+                                          );
+                                        });
+                                        return Column(
+                                          crossAxisAlignment: CrossAxisAlignment.start,
+                                          children: rows,
+                                        );
+                                      }
+                                    },
+                                  ),
+                                  const SizedBox(height: 20),
+                                  Text(
+                                    'Other Topics',
+                                    style: GoogleFonts.nunito(color: Colors.black, fontSize: 28),
+                                  ),
+                                  const SizedBox(height: 20),
+                                  
+                                  FutureBuilder<List<String>>(
+                                    future: Future.value(otherTopics), // Assuming otherTopics is a list of other topics
+                                    builder: (context, snapshot) {
+                                      if (snapshot.connectionState == ConnectionState.waiting) {
+                                        return Center(child: CircularProgressIndicator());
+                                      } else if (snapshot.hasError) {
+                                        return Center(child: Text('Error loading topics'));
+                                      } else {
+                                        List<String> topics = snapshot.data ?? [];
+                                        int numTopics = topics.length;
+                                        int numTopicsPerRow = 4; // Adjust the number of topics per row as needed
+                                        int numRows = (numTopics / numTopicsPerRow).ceil();
+                                        List<Widget> rows = List.generate(numRows, (rowIndex) {
+                                          List<Widget> rowChildren = [];
+                                          for (int i = 0; i < numTopicsPerRow; i++) {
+                                            int index = rowIndex * numTopicsPerRow + i;
+                                            const SizedBox(height: 10);
+                                            if (index < numTopics) {
+                                              rowChildren.add(
+                                                Flexible(
+                                                  child: Padding(
+                                                    padding: const EdgeInsets.symmetric(horizontal: 10.0, vertical: 10),
+                                                    child: Container(
+                                                      height: 200,
+                                                      decoration: BoxDecoration(
+                                                        borderRadius: BorderRadius.circular(10),
+                                                        color: Colors.white,
+                                                        boxShadow: [
+                                                          BoxShadow(
+                                                            color: Colors.black.withOpacity(0.2),
+                                                            spreadRadius: 1,
+                                                            blurRadius: 2,
+                                                            offset: const Offset(0, 1),
+                                                          ),
+                                                        ],
+                                                      ),
+                                                      child: InkWell(
+                                                        onTap: () async {
+                                                          print('Topic ${index + 1}: ${topics[index]} pressed');
+                                                          // Add functionality here if needed
+                                                        },
+                                                        child: Center(
+                                                          child: Column(
+                                                            mainAxisAlignment: MainAxisAlignment.center,
+                                                            crossAxisAlignment: CrossAxisAlignment.center,
+                                                            children: [
+                                                              const Icon(
+                                                                Icons.topic, // Replace with appropriate icon
+                                                                size: 60,
+                                                                color: Colors.blue,
+                                                              ),
+                                                              const SizedBox(height: 10),
+                                                              Text(
+                                                                topics[index],
+                                                                style: const TextStyle(fontSize: 16),
+                                                                textAlign: TextAlign.center,
+                                                              ),
+                                                            ],
+                                                          ),
+                                                        ),
+                                                      ),
+                                                    ),
+                                                  ),
+                                                ),
+                                              );
+                                            } else {
+                                              rowChildren.add(Flexible(child: SizedBox())); // Add an empty Flexible widget for even distribution
+                                            }
+                                          }
+                                          return Row(
+                                            children: rowChildren,
+                                          );
+                                        });
+                                        return Column(
+                                          crossAxisAlignment: CrossAxisAlignment.start,
+                                          children: rows,
+                                        );
+                                      }
+                                    },
+                                  ),
+
+                                ],
+                              ),
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+                Expanded(
+                  flex: 1,
+                  child: Container(
+                    height: MediaQuery.of(context).size.height,
+                    width: MediaQuery.of(context).size.width * 1 / 3,
+                    margin: const EdgeInsets.fromLTRB(0, 30, 30, 30),
+                    child: Column(
+                      children: [
+                        Expanded(
+                          child: Container(
+                            padding: EdgeInsets.all(16.0),
+                            decoration: BoxDecoration(
+                              borderRadius: BorderRadius.circular(12),
+                              color: const Color(0xFFf3edf6).withOpacity(1),
+                              boxShadow: [
+                                BoxShadow(
+                                  color: Colors.black.withOpacity(0.2),
+                                  spreadRadius: 5,
+                                  blurRadius: 10,
+                                  offset: const Offset(0, 3),
+                                ),
+                              ],
+                            ),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.center,
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              children: [
+                                Text(
+                                  'XP Level',
+                                  style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+                                ),
+                                SizedBox(height: 10),
+                                Container(
+                                  width: double.infinity,
+                                  height: 40,
+                                  decoration: BoxDecoration(
+                                    border: Border.all(color: Colors.black),
+                                    borderRadius: BorderRadius.circular(20),
+                                  ),
+                                  child: Stack(
+                                    children: [
+                                      FractionallySizedBox(
+                                        widthFactor: xpLevel / 100,
+                                        child: Container(
+                                          decoration: BoxDecoration(
+                                            color: Colors.blue,
+                                            borderRadius: BorderRadius.circular(20),
+                                          ),
+                                          alignment: Alignment.center,
+                                          child: Text(
+                                            '$xpLevel XP - ${_getXPLevelDescription(xpLevel)}',
+                                            style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                                          ),
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                ),
+                                SizedBox(height: 10),
+                              ],
+                            ),
+                          ),
+                        ),
+                        SizedBox(height: 30),
+                        Expanded(
+                          flex: 7,
+                          child: Container(
+                            width: MediaQuery.of(context).size.width * 1 / 3,
+                            padding: EdgeInsets.all(16.0),
+                            decoration: BoxDecoration(
+                              borderRadius: BorderRadius.circular(12),
+                              color: const Color(0xFFf3edf6).withOpacity(1),
+                              boxShadow: [
+                                BoxShadow(
+                                  color: Colors.black.withOpacity(0.2),
+                                  spreadRadius: 5,
+                                  blurRadius: 10,
+                                  offset: const Offset(0, 3),
+                                ),
+                              ],
+                            ),
+                            child: SingleChildScrollView(
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                children: [
+                                  const SizedBox(height: 20),
+                                  Row(
+                                    children: [
+                                      Spacer(),
+                                      Text(
+                                        'Quiz History',
+                                        style: GoogleFonts.nunito(color: Colors.black, fontSize: 28),
+                                      ),
+                                      SizedBox(width: MediaQuery.of(context).size.width * 1/12), // Adjust the width as needed
+                                      ElevatedButton(
+                                        onPressed: () {
+                                          Navigator.push(
+                                            context,
+                                            MaterialPageRoute(
+                                              builder: (context) => QuizHistoryPage(),
+                                            ),
+                                          );
+                                        },
+                                        child: const Text(
+                                          'View All',
+                                          style: TextStyle(color: Colors.black),
+                                        ),
+                                      ),
+                                    ],
+                                  ),
+                                  const SizedBox(height: 20),
+                                  FutureBuilder<List<String>>(
+                                    future: getQuizNames(recentQuizzes),
+                                    builder: (context, snapshot) {
+                                      if (snapshot.connectionState == ConnectionState.waiting) {
+                                        return Center(child: CircularProgressIndicator());
+                                      } else if (snapshot.hasError) {
+                                        return Center(child: Text('Error loading quiz names'));
+                                      } else {
+                                        List<String> quizNames = snapshot.data ?? [];
+                                        int numRecentQuizzes = quizNames.length;
+                                        int numQuizzesPerRow = 2;
+                                        int numRows = (numRecentQuizzes / numQuizzesPerRow).ceil();
+                                        List<Widget> rows = List.generate(numRows, (rowIndex) {
+                                          List<Widget> rowChildren = [];
+                                          for (int i = 0; i < numQuizzesPerRow; i++) {
+                                            int index = rowIndex * numQuizzesPerRow + i;
+                                            const SizedBox(height: 10);
+                                            if (index < numRecentQuizzes) {
+                                              rowChildren.add(
+                                                Expanded(
+                                                  child: Padding(
+                                                    padding: const EdgeInsets.symmetric(horizontal: 10.0, vertical: 10),
+                                                    child: Container(
+                                                      height: 200,
+                                                      decoration: BoxDecoration(
+                                                        borderRadius: BorderRadius.circular(10),
+                                                        color: Colors.white,
+                                                        boxShadow: [
+                                                          BoxShadow(
+                                                            color: Colors.black.withOpacity(0.2),
+                                                            spreadRadius: 1,
+                                                            blurRadius: 2,
+                                                            offset: const Offset(0, 1),
+                                                          ),
+                                                        ],
+                                                      ),
+                                                      child: InkWell(
+                                                        onTap: () async {
+                                                          await _getloadedQuestions(recentQuizzes[index]);
+                                                          await _loadQuizAttemptData(recentQuizzes[index]);
+                                                          _quizSummaryButton(loadedQuestions, quizAttemptData);
+                                                        },
+                                                        child: Center(
+                                                          child: Column(
+                                                            mainAxisAlignment: MainAxisAlignment.center,
+                                                            crossAxisAlignment: CrossAxisAlignment.center,
+                                                            children: [
+                                                              const Icon(
+                                                                Icons.history,
+                                                                size: 60,
+                                                                color: Colors.blue,
+                                                              ),
+                                                              const SizedBox(height: 10),
+                                                              Text(
+                                                                'Recent Quiz ${index + 1}: ${quizNames[index]}',
+                                                                style: const TextStyle(fontSize: 16),
+                                                                textAlign: TextAlign.center,
+                                                              ),
+                                                            ],
+                                                          ),
+                                                        ),
+                                                      ),
+                                                    ),
+                                                  ),
+                                                ),
+                                              );
+                                            }
+                                          }
+                                          return Row(
+                                            children: rowChildren,
+                                          );
+                                        });
+                                        return Column(
+                                          crossAxisAlignment: CrossAxisAlignment.start,
+                                          children: rows,
+                                        );
+                                      }
+                                    },
+                                  ),
+                                  const SizedBox(height: 20),
+                                ],
+                              ),
+                            ),
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+              ],
+            )
+          : Center(
+              child: ElevatedButton(
+                onPressed: () {
+                  // Navigate to login page
+                },
+                child: Text('Login'),
+              ),
+            ),
     );
   }
+
+
+  String _getXPLevelDescription(int xp) {
+    if (xp >= 0 && xp <= 20) {
+      return 'Beginner';
+    } else if (xp >= 21 && xp <= 40) {
+      return 'Intermediate';
+    } else if (xp >= 41 && xp <= 60) {
+      return 'Advanced';
+    } else if (xp >= 61 && xp <= 80) {
+      return 'Expert';
+    } else {
+      return 'Master';
+    }
+  }
 }
-
-
-
-
-
